requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-admin-display==1.3.0
django-admin-index==3.1.1
django-adminactions==2.1.0
django-adminplus==0.6
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-defender==0.9.7
django-grappelli==3.0.8
django-import-export==3.3.8
django-jazzmin==3.0.0
django-modeltranslation==0.18.11
django-optimistic-lock==1.0.0
django-ordered-model==3.7.4
django-redis==5.4.0
django-rest-framework==0.1.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
django-tinymce==4.0.0
django_constant==1.0.0
djangoql==0.18.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
graphql-core==3.2.3
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kombu==5.3.7
MarkupPy==1.14
msgpack==1.0.8
mypy==1.7.1
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.3.0
ply==3.11
priority==1.3.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
strawberry-django-auth==0.376.11
strawberry-graphql==0.227.1
strawberry-graphql-django==0.38.0
tablib==3.5.0
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.4.post2
